Massive reorganization + adding lots of fasteners.
I've tried to organize almost everything in functional subassemblies, as
well as moving the actual part/assembly files into appropriate folders:
there's a folder for custom parts I'm machining, and a folder for
off-the-shelf (even if we modify them) parts.

The biggest downside to the subassembly approach is that it seems to put
the kibosh on moving components from the top-level assembly., whereas
before we could simulate the range of motion of the X-Y axes, etc.  I
miss it, but it's not the biggest thing.  I prefer the cleaner file
structure.

We've also added fasteners for everything EXCEPT for the
rail-to-extrusion attachment.  There's so many, and I got lazy for the
day. :)

Next up is adding the remaining fasteners, and then starting to ensure
the proper material is configured for every part so we can start to
estimate the finished weight, and after that, the cost of individual
parts best as we can can determine from existing pricing data or, in the
case of custom parts, best guesses at quotes.