Add in the first cut off software for the vacuum controller.
We have some bog-standard ANSI C targetting the AVR library to control
the ATtiny and blip the valves and all of that.

Right now, it's configured that it recognizes the momentary blips from
the pedal, and does the following:
  - if the vacuum is not applied, it applies it and leaves it on
  - if the vacuum is applied, it turns it off and briefly opens the air
    valve to help provide some positive pressure to clear the line
    faster, thereby leaving the component on the board and not stuck to
    the nozzle

We're also tightly integrated with simavr, which is a software simulator
for running Atmel AVR code.  We utilize a lot of its provided build
  scripts and helpers to build the code and instrument it.  It's
  currently set up that we build a CLI program which runs the core and
  lets us send button presses to emulate the external trigger of the
  pedal, all while tracing the main pins and outputting that data to a
  gtkwave-compatible file.